README: make badge link to add-ons marketplace
Some users, like myself, expect the big bright button at the top of
README.md to lead to the installation.  Right now, the first badge at
the top of README.md isn't linking anywhere, but on GitHub it just leads
to camo.githubusercontent.com.

Turn the "mozilla add-on" badge in the README.md file into a clickable
link leading to the Firefox Add-ons Marketplace.